# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_readygeneration_readygeneration-backend_internal_domain.CompetitionLevel:
    enum:
    - low
    - medium
    - high
    - very_high
    type: string
    x-enum-varnames:
    - CompetitionLow
    - CompetitionMedium
    - CompetitionHigh
    - CompetitionVeryHigh
  github_com_readygeneration_readygeneration-backend_internal_domain.DifficultyLevel:
    enum:
    - easy
    - medium
    - hard
    - very_hard
    type: string
    x-enum-varnames:
    - DifficultyEasy
    - DifficultyMedium
    - DifficultyHard
    - DifficultyVeryHard
  github_com_readygeneration_readygeneration-backend_internal_domain.FunderType:
    enum:
    - federal
    - state
    - local
    - private
    - foundation
    - corporate
    type: string
    x-enum-varnames:
    - FunderFederal
    - FunderState
    - FunderLocal
    - FunderPrivate
    - FunderFoundation
    - FunderCorporate
  github_com_readygeneration_readygeneration-backend_internal_domain.GrantStatus:
    enum:
    - active
    - closed
    - archived
    - draft
    type: string
    x-enum-varnames:
    - GrantStatusActive
    - GrantStatusClosed
    - GrantStatusArchived
    - GrantStatusDraft
  github_com_readygeneration_readygeneration-backend_pkg_response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.APIError'
      meta:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Meta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handler.createGrantRequest:
    properties:
      agency:
        type: string
      application_url:
        type: string
      category:
        type: string
      competition_level:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_internal_domain.CompetitionLevel'
      deadline:
        type: string
      description:
        type: string
      difficulty_level:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_internal_domain.DifficultyLevel'
      eligible_org_types:
        items:
          type: string
        type: array
      eligible_populations:
        items:
          type: string
        type: array
      eligible_states:
        items:
          type: string
        type: array
      focus_areas:
        items:
          type: string
        type: array
      funder_name:
        type: string
      funder_type:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_internal_domain.FunderType'
      match_percentage:
        type: number
      max_award_amount:
        type: integer
      min_award_amount:
        type: integer
      open_date:
        type: string
      requires_501c3:
        type: boolean
      requires_audited_fin:
        type: boolean
      requires_match:
        type: boolean
      slug:
        type: string
      status:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_internal_domain.GrantStatus'
      synopsis:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      total_funding_available:
        type: integer
    required:
    - funder_name
    - funder_type
    - slug
    - title
    type: object
...
  internal_handler.ingestNOFORequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
...
  internal_handler.updateGrantRequest:
    properties:
      deadline:
        type: string
      description:
        type: string
      status:
        type: string
      synopsis:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Grant intelligence platform API
  title: ReadyGeneration API
  version: "1.0"
paths:
  /admin/grants:
    post:
      consumes:
      - application/json
      parameters:
      - description: Grant payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createGrantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create a new grant (admin)
      tags:
      - grants
  /admin/grants/{id}:
    delete:
      parameters:
      - description: Grant UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Archive a grant (admin)
      tags:
      - grants
    patch:
      consumes:
      - application/json
      parameters:
      - description: Grant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateGrantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update a grant's mutable fields
      tags:
      - grants
  /admin/grants/{id}/nofo:
    post:
      consumes:
      - application/json
      parameters:
      - description: Grant UUID
        in: path
        name: id
        required: true
        type: string
      - description: NOFO text
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ingestNOFORequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Ingest NOFO document text for a grant (admin)
      tags:
      - grants
...
  /grants:
    get:
      parameters:
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      - description: Comma-separated statuses
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: List active grants
      tags:
      - grants
  /grants/{id}:
    get:
      parameters:
      - description: Grant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Get a grant by ID
      tags:
      - grants
  /grants/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Search grants by title keyword
      tags:
      - grants
  /grants/semantic-search:
    get:
      parameters:
      - description: Free-text description
        in: query
        name: q
        required: true
        type: string
      - description: Number of results (default 10)
        in: query
        name: top_k
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Semantic (vector) search for grants
      tags:
      - grants
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"